Safari-609.1.11.1/LayoutTests/js/dfg-use-function-as-variable-not-constant-expected.txt

Tests what happens when you use a function as a variable, and you put things into it that aren't constants.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS run_tests(13) is 26
PASS successfullyParsed is true

TEST COMPLETE
